AI_Server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
transformers==4.47.1
torch==2.5.1
torchvision==0.20.1
pillow==11.0.0
opencv-python==4.10.0.84
numpy==1.26.4
